+++ tests/test_io.py
@@ -1,6 +1,7 @@
 import os
 import pytest
 import numpy as np
+from copy import deepcopy as dc
 from numpy.testing import assert_allclose
 
 from emg3d import meshes, models, fields, utils, io
@@ -142,22 +143,22 @@ def test_save_and_load(tmpdir, capsys):
     # Save it.
     io.save(tmpdir+'/test', emg3d=grid, discretize=grid2, model=model,
             broken=grid3, a=None,
-            field=field, what={'f': field.fx, 12: 12}, backend="numpy")
+            field=field, what={'f': field.fx, 12: 12}, backend="npz")
     outstr, _ = capsys.readouterr()
     assert 'WARNING :: Could not serialize <broken>' in outstr
 
     # Load it.
-    out = io.load(str(tmpdir+'/test.npz'), allow_pickle=True)
+    out_npz = io.load(str(tmpdir+'/test.npz'), allow_pickle=True)
     outstr, _ = capsys.readouterr()
     assert 'Loaded file' in outstr
     assert 'test.npz' in outstr
     assert utils.__version__ in outstr
 
-    assert out['Model']['model'] == model
-    assert_allclose(field.fx, out['Field']['field'].fx)
-    assert_allclose(grid.vol, out['TensorMesh']['emg3d'].vol)
-    assert_allclose(grid.vol, out['TensorMesh']['discretize'].vol)
-    assert_allclose(out['Data']['what']['f'], field.fx)
+    assert out_npz['Model']['model'] == model
+    assert_allclose(field.fx, out_npz['Field']['field'].fx)
+    assert_allclose(grid.vol, out_npz['TensorMesh']['emg3d'].vol)
+    assert_allclose(grid.vol, out_npz['TensorMesh']['discretize'].vol)
+    assert_allclose(out_npz['Data']['what']['f'], field.fx)
 
     # Check message from loading another file
 
@@ -181,18 +182,86 @@ def test_save_and_load(tmpdir, capsys):
     with pytest.raises(NotImplementedError):
         io.load('ttt.abc')
 
+    # Ensure deprecated backend/extension still work.
+    io.save(tmpdir+'/ttt', backend='numpy')
+
     # Test h5py.
     if h5py:
         io.save(tmpdir+'/test', emg3d=grid, discretize=grid2,
+                a=1.0, b=1+1j,
                 model=model, field=field, what={'f': field.fx})
         out_h5 = io.load(str(tmpdir+'/test.h5'))
         assert out_h5['Model']['model'] == model
+        assert out_h5['Data']['a'] == 1.0
+        assert out_h5['Data']['b'] == 1+1j
         assert_allclose(field.fx, out_h5['Field']['field'].fx)
         assert_allclose(grid.vol, out_h5['TensorMesh']['emg3d'].vol)
         assert_allclose(grid.vol, out_h5['TensorMesh']['discretize'].vol)
         assert_allclose(out_h5['Data']['what']['f'], field.fx)
+
+        assert io._compare_dicts(out_h5, out_npz) is True
     else:
+        with pytest.raises(ImportError):
+            # Ensure deprecated backend/extension still work.
+            io.save(tmpdir+'/ttt', backend='h5py')
         with pytest.raises(ImportError):
             io.save(tmpdir+'/test-h5', grid=grid)
         with pytest.raises(ImportError):
             io.load(str(tmpdir+'/test-h5.h5'))
+
+    # Test json.
+    io.save(tmpdir+'/test', emg3d=grid, discretize=grid2,
+            a=1.0, b=1+1j,
+            model=model, field=field, what={'f': field.fx}, backend='json')
+    out_json = io.load(str(tmpdir+'/test.json'))
+    assert out_json['Model']['model'] == model
+    assert out_json['Data']['a'] == 1.0
+    assert out_json['Data']['b'] == 1+1j
+    assert_allclose(field.fx, out_json['Field']['field'].fx)
+    assert_allclose(grid.vol, out_json['TensorMesh']['emg3d'].vol)
+    assert_allclose(grid.vol, out_json['TensorMesh']['discretize'].vol)
+    assert_allclose(out_json['Data']['what']['f'], field.fx)
+
+    assert io._compare_dicts(out_json, out_npz) is True
+
+
+def test_compare_dicts(capsys):
+    # Create test data
+    grid = utils.TensorMesh(
+            [np.array([100, 4]), np.array([100, 8]), np.array([100, 16])],
+            np.zeros(3))
+
+    model = utils.Model(grid, res_x=1., res_y=2., res_z=3., mu_r=4.)
+
+    e1 = create_dummy(*grid.vnEx)
+    e2 = create_dummy(*grid.vnEy)
+    e3 = create_dummy(*grid.vnEz)
+    ee = utils.Field(e1, e2, e3, freq=.938)
+
+    dict1 = io._dict_serialize(
+            {'model': model, 'grid': grid, 'field': ee,
+             'a': 1, 'b': None, 'c': 1e-9+1j*1e13,
+             'd': {'aa': np.arange(10), 'bb': np.sqrt(1.0),
+                   'cc': {'another': 1}, 'dd': None}
+             })
+
+    dict2 = dc(dict1)
+    out = io._compare_dicts(dict1, dict2)
+    assert out is True
+
+    del dict1['Data']['d']['bb']
+    del dict2['Field']
+    del dict2['Model']['model']['mu_r']
+    dict2['TensorMesh']['grid']['hy'] *= 2
+    dict2['whatever'] = 'whatever'
+    dict2['2onlydict'] = {'booh': 12}
+
+    out = io._compare_dicts(dict1, dict2, True)
+    assert out is False
+    outstr, _ = capsys.readouterr()
+    assert " True  :: Model      > model      > res_x" in outstr
+    assert "  {1}  ::                           mu_r" in outstr
+    assert " False ::                           hy" in outstr
+    assert " True  ::                           cc         > another" in outstr
+    assert "  {2}  :: Data       > d          > bb" in outstr
+    assert "  {2}  :: 2onlydict" in outstr